relocate array reversion
